+++ src/v3/mod.rs
@@ -8,18 +8,16 @@ mod dispatcher;
 mod handshake;
 mod publish;
 mod router;
-mod selector;
 mod server;
 mod shared;
 mod sink;
 
 pub type Session<St> = crate::Session<MqttSink, St>;
 
-pub use self::control::{ControlMessage, ControlResult};
+pub use self::control::{Control, ControlAck};
 pub use self::handshake::{Handshake, HandshakeAck};
 pub use self::publish::Publish;
 pub use self::router::Router;
-pub use self::selector::Selector;
 pub use self::server::MqttServer;
 pub use self::sink::{MqttSink, PublishBuilder, SubscribeBuilder, UnsubscribeBuilder};